AttachedFileDownloader calls Aws::S3 directly
Because of the way ActiveStorage is configured in test mode, it was
impossible to get decent test converage for failures that occur during
the download process.

So, this uses the `get_object` command of the AWS S3 ruby gem directly
instead of going through ActiveStorage, so we're able to use its
built-in stubbing support.