Definition invalid when management charge references optional field
If the field used for flat-rate or column-based management charge
calculation is optional, then the user can submit rows within their
return that will pass validation, but fail the management charge
calculation.

To prevent this we check that these fields are required within the
transpiler, and raise a `Transpiler::Error` when this occurs.

The existing `field_by_name` method uses the data warehouse field name,
whereas the management charge calculation uses the sheet name, hence the
need for adding `field_by_sheet_name` to the presenter.

Both RM3710 and RM858 failed this check, so they have been corrected.