Raise a Transpiler::Error when known field missing
Previously this was a DataWarehouse::KnownFields::NotFoundError.
This wouldn't bubble up as an acceptable transpiler problem and be
mapped into a validation failure on the FDL; it would just crash with
a Rails Red Screen.

Rather than catch the NotFoundError, be proactive and add an
AST::Field#error method, raising a Transpiler::Error with the string
return value of this method. If we need further intra-field errors in
the future we can add them here rather than filling the Transpiler full
of `rescue` blocks.