Stricter validations on Task
When Task was originally created, it was anticipated that it might be used for tasks that aren't just submissions. This has not proven to be true, and we are about to allow admins create tasks, so I'm adding some stricter validations to tasks.

Period year, period month, due on and framework should all be mandatory. There are no instances on production where this isn't the case, so no migration needed.

I've also added validation to make sure there aren't duplicate tasks for the same period for a supplier on a framework.