Expect HTTP Authorization header instead of 'X-Auth-Id'
Instead of using the 'X-Auth-Id' header to determine the user that is
accessing the API, use HTTP Authorization and expect the bearer to be
a JSON web token signed with Auth0's private key.

The public key used to verify the JSON web token is stored in the
`AUTH0_JWT_PUBLIC_KEY` environment variable.

When testing, we use our own key pair to sign the JSON web token used
in the request. It duplicates the payload attributes that we use and
also sets the expiry time to 10 hours in the future, to match Auth0's
default.