Introduce an AST::SemanticChecker
Rather than trying to spot any error on first use as we transpile,
make this a step that the transpiler invokes before it starts generating
classes (it probably shouldn't be the transpiler's responsibility; we
could invoke AST::SemanticChecker#run from the Generator, but that would
make this commit a good bit larger).

Move the checks that the Transpiler was doing – management charge and
field errors into the AST::SemanticChecker. You can argue that probably
now fields shouldn't be checking themselves, but there's limited time
to add even this.

Also, add dependent reference checking: if a field depends_on another
field that doesn't exist, say so.