Expose submission's file key and name to API
For GPAAS, we can no longer use a temporary file URL to share a submission file with a supplier. The frontend will therefore need to know the file key to be able to use the AWS API to retrieve the file and give it to the user.

I did a slight refactor of SerializableSubmission spec, so it uses a completed submission factory, rather than a no business submission so we can better test the file key param.