Tests use their own version of RM3786
The data warehouse export tests are based heavily on RM3786, as it was
one of the first frameworks we onboarded. Now this framework has been
changed, the tests were failing.

I've fixed them by creating a framework in the test suite, based on the
original version of RM3786. I've renamed it to "RM3786v1" because FDL
files are currently prioritised over database entries, which means the
version loaded from the fixture file would be ignored.