(fix) Add codemirror to the yarn.lock file
This dependency was missing from `yarn.lock`, which means that when
Docker runs `yarn` it does not get installed. The absence of this
library means many of the tests fail.

Running `yarn` via `docker exec` installs the missing dependency,
updates the lock file and this fixes the failing tests.

I have verified that this works in a newly clone repo by deleting my
`node_modules` directory, stopping and destroying all running
containers, and running the build from a clean state.