Fix FDL double-sanitisation issue
Framework::Definition#[] is responsible for getting a definition from
the cache. Framework::Language#[] is responsible for getting the source
from either the DB or the filesystem and transpiling it.

In the circumstance that a framework was in the database but had a
period in the short_name, a double sanitisation issue occurred, meaning
that Language#[] would try and look for (say) 'RM999_5' in the database
when 'RM999.5' was there, resulting in

```
Framework::Definition::MissingError: There is no framework definition for "RM999_5"
```

Take the responsibility for sanitisation away from Definition and leave
it with Language to avoid this.