(chore) Add FDL files to spec/fixtures
Some of the framework definition files in app/models/framework/defintion
are used in the test suite, and this is a blocker on converting them
into first-class frameworks in the database.

To sever this dependency, this commit copies those FDL files that are
used in the test suite into the spec/fixtures directory.