(fix) Don't use gzip when importing database backups
The instructions for getting a production database dump don't result in
one receiving a gzipped archive, they just result in a `.tar`. The
instructions tell you to `gzip` the file so the importer can handle it,
but all the importer does it `gunzip` it before piping it into
`pg_restore`.

This change therefore makes the `db:restore` task process `.tar` files,
not gzipped ones. Now that the `pg_restore` command is a single process,
we can also use the safer multi-argument form of `system` to pass
parameters to `pg_restore` without worrying about shell string
interpolation/injection.