(fix) Align development Python version with that of production
When investigating bug 1112 ("Known fields are treated as incorrect
types in certain scenarios") I was unable to reproduce some of the
failing scenarios in my Docker development environment.

After confirming that the bug was caused by the Python-based `in2csv`
tool that we use in the ingest, I noticed that the behaviour of this
tool changes depending on the Python version that it is used with.

In production, we are using Python 3.7.2 (specified by `runtime.txt`)
and in development we are using `apt-get` to install 2.x.y (specified in
the `Dockerfile`).

Hence, ensure that we install the same version of Python in development
as in production. These will still need to be kept in sync manually. (I
didn't think it was worth spending time now to make the Dockerfile use
the `runtime.txt` file since I think we'll be changing the Python
version very infrequently.) This means having to install Python from
source instead of apt-get, since the latter only seems to have the
latest version of Python 3.7 (at time of writing, this was 3.7.4).

It's worth noting that we also use apt-get to install Python in Travis.
We should consider finding a way to lock that version, too.