(fix) Disable in2csv's type inference, as it can mis-infer
By default, the `in2csv` tool used by `Ingest::Converter` performs some
inference to guess the type of a column. For example, given a column
with mixed values like "True" and "Y" and "1", it will decide that this
is a column of booleans, and in its output it will normalise them to
"True" and "False".

The full list of types which in2csv will infer can be found
[here](https://github.com/wireservice/csvkit/blob/ee9d9eb5922619b97c4d6b1e4132c79d415fe235/csvkit/cli.py#L300).
To understand how the rules about how a given type is inferred, take a
look at the implementation of the `cast` method for the [different subclasses
of `agate.DataType`](https://agate.readthedocs.io/en/1.6.1/api/data_types.html#supported-types).
(csvkit appears to be a thin wrapper around the "agate" library.)

Sometimes, the inference means that in2csv will unexpectedly convert a
numerical column into "True" or "False", or a string like "1 Year" into
a date like "0002-01-01". This is an undesirable behaviour and it is
preventing users from successfully submitting returns. in2csv does not
provide many configuration options relating to type inference, so I
think our only option is to disable it completely using the
`--no-inference` flag.

The obvious question is then whether we lose anything by disabling
inference. It _seems_ to me that we do not. The existing test cases
continue to pass. For many of the types for which in2csv provides
inference, we do not need it, since it seems that our templates mandate
a fixed format anyway. For example, "booleans" (`YesNo` in FDL) should
always be filled in by the user as 'Y' or 'N', and dates as DD/MM/YYYY.

There remains the possibility that disabling inference does cause our
ingestion to be less permissive than it was before. CCS pointed out that
is better for the user to be faced with a validation error that they can
fix by modifying their data than an error that they are powerless to
solve.

It is also worth noting that _some_ of the example spreadsheets did not
exhibit the issue when tested with Python 2, which is what we were using
in our development environment until recently. I do not think it is
worth spending time to understand the reason for this.

The new test fixtures here come from filling the relevant columns of the
real template with data from the real example spreadsheets.