(chore) Load framework definitions from spec/fixtures
A few places in the test suite are relying on the
Framework::Definition[] method to load FDL from
app/models/framework/definition.

Here we replicate the functionality of this class in FrameworkLoader,
which loads FDL from spec/fixtures instead, and use this class in place
of Framework::Definition in the tests.