(chore) Copy filesystem FDL into the database
This migration finds all the Frameworks that do not have a
definition_source in the database, and populates this field from the FDL
on disk.

I have verified that all Frameworks we currently have in production are
matched to a FDL in the filesystem, and all FDL currently in the
filesystem is imported by this migration.