(refactor) Remove memoization of framework definition in SubmissionEntry#framework_definition
In preparation for allowing users to edit definitions of published
frameworks, I was auditing all the places where we store framework
definitions, and noticed this.

It is safe to remove the memoization, because the method is actually
unused!

Further investigation revealed that this method is duplicated by
Ingest::Loader#add_validation_errors!, so get rid of the latter method,
and replace it with the former (which has good unit tests).