(refactor) Do not cache framework definitions globally
The original intention of definition caching (according to 1a69f8b) was
to prevent us from generating the anonymous framework classes too
frequently.

This remains a desirable property of the system, but now that we want to
add the ability to edit frameworks, we cannot keep the global in-memory
cache (unless we want to come up with some invalidation mechanism that
works cross-process/machine).

Rely on the following mechanisms to avoid overzealous recompilation of
frameworks whilst giving us control over the lifetime of the cache:

- The existing memoization in Framework#definition. Added a test to
confirm this behaviour.

- A new ephemeral cache which is shared between rows when generating the
invoice and contract data warehouse exports. It is worth noting that by
removing the global cache, we now introduce {number of frameworks}
additional queries to each export job, to load the definition of each
framework from the database. After a quick think, it was not obvious how
to adapt our current export mechanism to avoid this, so I have not
addressed it. Given the relative infrequency of the export jobs and
since at the time of writing we only have around 70 frameworks, I did
not consider it to be worth spending too much time on.