(feature) Reject submission if the framework definition changes mid-ingest
Now that we wish to allow an admin user to edit a published framework,
there is a (slim) chance that the framework definition will change
between the moments that

1. a submission is validated by SubmissionIngestionJob, and that
2. the management charge is calculated by SubmissionManagementChargeCalculationJob.

If this occurs, we do not wish to proceed with the management charge
calculation; the user must submit another return. Detect when this
occurs, and add a new management_charge_calculation_failed submission
state to reflect it.

Log the error to Rollbar so that we can get a sense of how frequently
this happens in reality.