(feature) Allow multiple depends_on fields in the grammar
We'd like to allow fields to depend on multiple other fields; whereas
the following syntax is currently allowed:

    MyField from 'a field' depends_on 'another field' {
      'value 1' -> Lookup1
      'value 2' -> Lookup2
    }

We now want to allow multiple fields in the depends_on clause, allowing
matching over combinations of values from each column:

    MyField from 'a field' depends_on 'another field', 'final field' {
      'left 1', 'right 1' -> Lookup1
      'left 2', 'right 2' -> Lookup2
    }

This specifies that if 'another field' has the value 'left 1' and 'final
field' has the value 'right 1', then the valid values for 'a field' are
found in Lookup1, and similarly for the second match arm.

Here we change only the grammer so that lists of field names are allowed
to follow depends_on, and the matching dictionary can have lists of
strings as its keys. We also pluralise some of the internal method names
to reflect this, but no new semantics are added here.