(refactor) Don't include field names in dependent_field_inclusion.in
Imagine we have FDL like the following:

    InvoiceFields {
      A from 'field a'
      B from 'field b' depends on 'field a' {
        'one' -> Lookup1
        'two' -> Lookup2
      }
    }
    Lookups {
      Lookup1 [
        'Hi'
      ]
      Lookup2 [
        'There'
      ]
    }

Currently, this will generate the following dependent field inclusion
validation spec:

    {
      parent: 'field a',
      in: {
        'field a' => {
          'one' => ['Hi'],
          'two' => ['There']
        }
      }
    }

The use of 'field a' in the `in` section is redundant for performing the
validation: the `parent` field tells us which field to get the values
from. This also raises a question about how to model depends_on when
multiple dependent fields are present. For example, if we have the FDL:

    InvoiceFields {
      A from 'field a'
      B from 'field b'
      C from 'field c' depends on 'field a', 'field b' {
        'one', 'apple' -> Lookup1
        'two', 'orange' -> Lookup2
      }
    }
    Lookups {
      Lookup1 [
        'Hi'
      ]
      Lookup2 [
        'There'
      ]
    }

How would we represent the lookup table? Working within the current
design, the best approach I can see is:

    {
      parents: ['field a', 'field b'],
      in: {
        'field a' => {
          'one' => {
            'field b' => {
              'apple' => ['Hi']
            }
          },
          'two' => {
            'field b' => {
              'orange' => ['There']
            }
          }
        }
      }
    }

I think this is unwieldy, being both hard to generate and hard to
understand the resulting validator code. Instead, I would like this
structure:

    {
      parents: ['field a', 'field b'],
      in: {
        ['one', 'apple'] => ['Hi'],
        ['two', 'orange'] => ['There']
      }
    }

This mirrors the structure of the FDL and allows for simpler generation
and validation -- ruby allows arrays of strings as hash keys and these
can be used transparently alongside single-string keys. The validator
would look up the values of 'field a' and 'field b' in a row, and use
this pair of values as a key into the table of allowed values.

To work towards this design, I'm keeping the `parent` field singular,
as well as the keys of the `in` mapping, but I'm just stripping out the
layer containing the name of the dependent field.