(feature) Validate based on multiple dependent fields
After changing the grammar in a previous commit, we now update the
semantics so that if a field has multiple dependent fields listed in its
depends_on clause, the values of all thoese fields are used to determine
the set of valid values for the field.

This multi-field support is not special-cased; instead, the existing
behaviour of having a single dependent field is now modelled as a list
of dependent fields containing only a single item, so we're running the
same logic no matter how many dependent fields are used.

To perform the validation, we look up the values of all the dependent
fields in the current row, and then use this array of values as a key
into the table of allowed values that follow the `depends_on` clause. So
for example:

    InvoiceFields {
      A from 'field a'
      B from 'field b'
      C from 'field c' depends_on 'field a', 'field b' {
        'one', 'apple' -> Set1
        'two', 'orange' -> Set2
      }
    }

This says that if 'field a' has the value 'one', and 'field b' the value
'apple', then the valid values for 'field c' are in Set1. If 'field a'
has value 'two' and 'field b' contains 'orange', then the valid values
are in Set2.