(refactor) Obey Rubocop's 'next' rule
In the following commit I will add a third statement to the `if` block
here. This trips Rubocop into insisting on a different layout for this
code, namely using a guard clause to skip the loop body rather than
enclosing it.

I'm committing this change separately just so that my next commit makes
it clear what I'm actually adding.