(feature) Wildcard matcher for multi-keyed depends_on
When one field depends on multiple other fields, it may be desirable to
use a wildcard in the match arms. For example, the value of the second
field may only become significant when the first field has a particular
value:

    InvoiceFields {
      A from 'field a'
      B from 'field b'
      C from 'field c' depends_on 'field a', 'field b' {
        'one', * -> Lookup1
        'two', 'apples' -> Lookup2Apples
        'two', 'oranges' -> Lookup2Oranges
        'two', 'lemons' -> Lookup2Lemons
      }
    }

This says that if 'field a' has the value 'one', then 'field c' can have
any value in Lookup1, regardless of the value of 'field b'. However if
'field a' has the value 'two', then the value in 'field b' affects the
lookup for 'field c'.

This makes is quicker to specify the lookup when 'field a' = 'one',
without listing out every possible value for 'field b'. The "any"
operator `*` must be used explicitly, so that there's no ambiguity
between an intentional wildcard match and an accidental omitted field.
This also means the * operator can be used in any column, not just the
last one.