(fix) Fix issues with production -> development database restore
- Fix permissions error that occurred when trying to run
20190820105032_make_framework_definition_source_not_null migration in
development against a database restored from a production dump.

- Fail the Rake task if pg_restore fails; we previously thought that all
the errors were ignorable but this error has revealed that's not true.
Add the necessary flags to prevent the role-related errors that _can_ be
ignored.

A little more information about how we filter out certain objects from
the database dump:

Running pg_restore with the --list flag causes it to not restore the
given dump, but to output a so-called Table of Contents file, describing
the dumped objects in a machine-readable format.

We can edit this file and then, when we use pg_restore to restore the
dump, we can pass it the modified ToC file using the -L flag, which
causes it to only restore the objects specified by the given ToC.

We take advantage of this to selectively restore objects by removing
lines from the ToC.

Examples of the lines that we wish to remove:

4; 3079 16417 EXTENSION - postgis
5581; 0 0 COMMENT - EXTENSION postgis
5296; 0 16717 TABLE DATA public spatial_ref_sys rdsadmin
1511; 1255 18011 FUNCTION public reassign_owned() R0SXAw77VjJCye6n
5295; 3466 22235 EVENT TRIGGER - reassign_owned rdsadmin

Examples of lines that we wish to keep:

226; 1259 20747 TABLE public framework_lots rdsbroker_e8e8d7d2_5890_4f9d_8c9c_da43a9918e4c_manager
5357; 2606 21798 CONSTRAINT public framework_lots framework_lots_pkey rdsbroker_e8e8d7d2_5890_4f9d_8c9c_da43a9918e4c_manager