(fix) Correct use of $@ in shell scripts
This puts the bash variable $@ in double quotes, so that shell words are
preserved correctly. For example, if one runs:

    ./bin/dspec -e 'name of test'

Then there are two argument strings passed to `dspec`: `-e`, and `name
of test`. The bash variable $@ contains these two strings. If it's used
unquoted, the effect is to insert them into a command as though they had
not been quoted originally, resulting in a command like this being run:

    docker exec ... rspec -e name of test

That means `rspec` is run with the arguments: `-e`, `name`, `of` and
`test`; `name` is used to match tests by name, and `of` and `test` are
assumed to be filenames, which results in an error.

Quoting the $@ variable effectively preserves the original quoting of
the shell words, so we get this instead:

    docker exec ... rspec -e 'name of test'

This means that invocations of these scripts that pass multi-word quoted
arguments pass those arguments through correctly.

I have also removed the variable $@ from a code path where the argument
array has been checked to be empty.