(refactor) Remove Google Tag Manager from template
Noticed this whilst auditing inline JavaScript in preparation for adding
Content-Security-Policy header.

The GOOGLE_TAG_MANAGER_ID environment variable is not set in staging or
production, so the <script> element never gets emitted. CCS confirmed
that they do not use any sort of analytics for the admin site. So this
is safe to remove. The commit history reveals nothing about why it was
added to the template, but I'd guess it was just a copy-and-paste.