(feature) Creating Users in Auth0 now tries to create first before handling duplication errors
* Adding as new commit since this work required an additional refactor to introduce a FindUserInAuth0 service
* There was a danger with the previous approach that 2 attempts to the sync the same user at the same time could result in a race condition where between calling `get_users` and `create_user` another task has made that user (perhaps directly in auth0). In this scenario the process would continue to fail as before.
* We have also observed that the `get_users` endpoint can be slow in returning results and have seen it take ~4 minutes before recently created users start being returned. This is problematic as lots of user uploading happens in quick succession at the start of the month.
* This change flips the flow around so that we always try the happy path first and attempt to create the user. Only when they doesn't succeed, we try the safer call of `get_users` which protects us from state changing in between to apply an update.
* The error class coming back from auth0 is unusually named `Unsupported` which refers to an unhandled error condition from the auth0 library.

If we were to get a 404 we would get a Auth0::NotFound but it seems they do not have a mapping for 409 https://github.com/auth0/ruby-auth0/blob/9cc3044f201c20d2c11deea060514aeeefbe9435/lib/auth0/exception.rb

The contents of this error class is also a little odd. It stores a blob of JSON from _something_ which we then have to convert to a hash. There is no handy and reliably way to say `if error.class == Auth0Conflict` so we have to analyse the contents. This is flaky, so I have asserted on 2 columns to ensure we don't accidentally rescue errors for other types of conflicts.