(refactor) Reduce memory usage of Ingest::Loader
Currently, we build up an in-memory array of SubmissionEntry instances
as we process each row of the file. One instance per row. These
instances consume a lot of RAM.

We need these instances for two reasons:

1. To validate the row's data. We do this as we process each row.
2. To store the row's data so that we can insert it into the database.
We do this after processing all rows.

This change achieves 2. with less RAM consumption by throwing away the
SubmissionEntry instance after performing 1., and storing the row's data
in a simple array of columns instead.

This does _not_ solve the fundamental problem of the ingest, which is
that the RAM consumption of the process can grow in an unbounded
fashion. It is merely a "sticking-plaster" fix.