(fix) Increase Sidekiq memory from 2048M to 8192M
Decrease the likelihood of ingest jobs failing with an out of memory
error (problem described in detail in docs/stuck-ingests.md).

I have already deployed this change to staging and production, by
running the CF/deploy-app.sh script locally, with this diff applied on
top of the develop and master branches respectively.

The new memory limit was picked pragmatically as a balance between
correctness and price - we want to be able to ingest the large file
referred to in that documentation, but given the current infrequency of
such a large file it is unlikely that we will need to be able to process
more than one of them in parallel.