(fix) Increase Sidekiq memory from 8192M to 16384M
We hoped that the increase to 8192M in commit 1fd67ee would stop the
would fix the problem of the "stuck" returns described in that commit.
However, we are still seeing this happening on production, when a
largish (15.5MB) return is uploaded whilst the aforementioned very large
(44MB) return was also being ingested. We have not yet witnessed the
44MB return succeeding.

Since this work is being performed as part of support, we cannot spend
much time investigating or working on this, so we have agreed with CCS
that for now we will try increasing the memory again to see if this
fixes things.

I have already deployed this change to staging and production, by
running the CF/deploy-app.sh script locally, with this diff applied on
top of the develop and master branches respectively.