Move ingest jobs into a separate 'ingest' queue
When the Sidekiq process runs out of memory, it's killed, taking any of
the jobs currently running on the instance with it. This meant that a
failed ingest process, could also cause other types of jobs to fail,
such as the daily data warehouse export.

To avoid that, we'll move ingest jobs into their own queue.