Deploy two separate instances of Sidekiq, one for each queue
Allow the 'default' and 'ingest' queues to be processed from their own
set of Sidekiq worker instances.

Additionally, I've split out memory, instance count and concurrency, as
the ingest and default queues have very different footprints.

NB: This also fixes the /instance/instances/ typo that was causing the
number of instances to always be 1.