Restore ingest Sidekiq instance's 4G disk quota
Bob increased the Sidekiq disk quota to 4G on Sep 10th, after we
started seeing out of disk space errors (we think possibly caused by
ingest):
https://dxw.slack.com/archives/CEB4W7GNB/p1568108084011700?thread_ts=1568103768.006900&cid=CEB4W7GNB

He must have done this as an ad-hoc change from his machine, though,
because it's not reflected in the Sidekiq manifests, so the disk quota
got decreased back down to 1GiB when we did a prod deploy today.

So, ensure that this does not happen again, by adding it to the
Sidekiq manifests.

Now that we have split the Sidekiq worker into two separate instances, I
have only applied the increased disk space to the ingest instance, since
we speculated that it was ingest that was responsible for the high disk
usage.

I have already deployed this change to staging and production, by
running the CF/deploy-app.sh script locally, with this diff applied on
top of the develop and master branches respectively.