management_charge is either retrieved from the db or calculated on the fly
To avoid having to backfill 74157++ submissions with their management charge
total (and probably crippling the prod server), I propose we calculate it only
when the record is accessed, and then store it on the model.

Thousands of submissions will probably never be viewed again so it's not worth
calculating it for all of them (maybe??).

I will calculate the chage for the currently troublesome entries (see
https://dxw.zendesk.com/agent/tickets/10151) only, and backfill them as necessary.