A script to fix the "bad" URNs which have been introduced into some SubmissionEntries
Zendesk https://dxw.zendesk.com/agent/tickets/10226

A list of SubmissionEntries with "bad" URNs has been supplied by Timur. Before
we correct the cause of these bad URNs we need to clean up the data. The URN is
stored in both a db field `customer_urn` and the `data` json blob on the
SubmissionEntry. This script will find the bad records and replace both the
`customer_urn` field and the entry in the jsonb.

Performance on the prod database isn't great and I'm a little concerned this may
be a slow script to run.