Backfill submissions.management_charge_total
Viewing the supplier view for a supplier with many submissions and a
large number of entries would bring the site to a crawl calculating
totals on the fly. A previous PR fixed this by pre-calculating
submissions.management_charge_total on first usage,
but this is still not ideal:

  https://github.com/dxw/DataSubmissionServiceAPI/pull/533

Backfill all of submissions.management_charge_total for completed and
in_review submissions.