Allow config of Postgres work_mem
Add this to config/database.yml and default to 16MB. Default for most PG
installations, including GPaaS, is 4MB.

We are heavily IOPS-bound on GPaaS whereas memory is not at such a
premium. Using a larger work_mem allows Postgres to choose query plans
that don't thrash the disk so much whenever a working set of data can
fit within work_mem.