Ensure only valid ISO8601 dates are parsed and converted
We allow dates that are in ISO8601 format to be ingested, but we
transform them in to dd/mm/yyyy format to be consistent with other
expected dates.

The issue we saw was that certain invalid ISO8601 dates (eg. 2020-02-30)
were attempting to be parsed because they matched the regex
`\d{4}-\d{2}-\d{2}` causing an ArgumentError.

We expect these dates to be left alone so that they are caught by the
`IngestedDateValidator` and surfaced to the user.