Validate multi-column management charge calcs
In the last feature, we forgot to do an end to end test that included the
validations. Currently, if we try to add a framework with a multi-column
management charge, it chokes, because
`raise_when_management_field_invalid` expects the modifier to always be a
string. This updates the checker to cope with arrays too. I considered
always casting the modifier to an array, and then flattening it, but I think
checking the type and looping over if it's an array is less obtuse.