Add missing indexes to speed up daily export
The indexes on the 3 fields have been defined as BRIN
(Block Range Index) as these are much smaller than B-Tree
when indexing timestamps. See:

https://karolgalanciak.com/blog/2018/08/19/indexes-on-rails-how-to-make-the-most-of-your-postgres-database/

This work has been requested in order to improve the
problematic daily export procedure invoked via
`DataWarehouseExport.generate!` which runs out of memory
and times out regularly.

We believe that the 2 problem extractions are:

- `Export::Invoices::Extract`
- `Export::Contracts::Extract`

and that improvements to these may be brought about by
adding the following indexes:

- `tasks.updated_at`
- `submissions.updated_at`
- `submission_entries.updated_at`

We apply the indexes concurrently to avoid locking the
database tables. This requires:

- the `disable_ddl_transaction!` directive in the migration
- the application of increased `maintenance_work_mem` as per
  `1bd0dc6033ab34ea30ff8bd15bac492a8563992c`

Zendesk
: https://dxw.zendesk.com/agent/tickets/10463

Trello
: https://trello.com/c/GeaLyAWy/1162