Add example of multi-column lookup
It's helpful to have an example of a real multi-column
management charge lookup from a real framework definition.

In this example we see three options. Each key is composed
of two column names in a list (a literal array).
Importantly, we note that the double quotes are escaped.

i.e in order to retrieve a given percentage we will need to
address the lookup with an escaped key e.g.

 `"[\"1\", \"transactional\"]"`