Fix multi-column management charge calculation
This change is an attempt at fixing some Rollbar errors which were
triggered because the management charge lookup failed to find an value
in the percentages lookup table for a particular submission entry.

I decided to investigate the contract for passing the percentages lookup
to the Framework::ManagementChargeCalculator::ColumnBased class. By
putting a breakpoint in that class's initializer and running
specs/models/framework/definition/language/generator_spec.rb:395, I
noticed that the model spec for the management charge calculator does
not accurately reflect how the class is being used in the app. In the
app, the value_to_percentage argument for a multi-column management
charge is a hash with array-valued keys, whilst in the management charge
calculator's model specs we are passing a nested hash.

So, the model specs for the calculator were telling us nothing about
whether multi-column management charge calculation would actually work
in the app.

This commit updates the model spec so that it is reflective of how the
code is actually being used, and fixes the calculator so that it works
with this usage.

I was wondering why this mismatch of API contracts had not been caught
by any higher-level tests. By commenting out all of the management
charge calculators' percentage calculation methods and running the specs
to see what broke, I realised that we only have a couple of high-level
specs that test management charge calculation end-to-end. This does not
give me much confidence in the correct functioning of management charge
calculation. I think we should take some time, when we have it, to look
at adding some high-level specs for each type of management charge
calculation.

I am _still_ not fully confident that multiple-column management charge
calculation is now working. I have not tested it locally with an example
return, since the development environment is broken (trying to load the
admin or supplier web interfaces in a browser gives a "server
unexpectedly dropped the connection" error). We will need to sort this
out, but it won't be a five-minute thing since it looks like it's
Docker networking-related. I have agreed with Timur that he will do some
testing of the feature on staging, to give confidence.