(fix) Restore column order for the submissions export
Some recent changes to the column order and headings have broken CCS's
internal processes that consume these CSVs. To fix it, we need to name
TotalSpend back to InvoiceValue. The old InvoiceEntryCount is no longer
exported, but we need a column before InvoiceValue to put InvoiceValue
at the correct offset; the client has advised we can insert a column
with no values to achieve this.