Bulk user import runs as a background job
When bulk user imports contain large numbers of users the service times
out. In order to mitigate this, bulk user imports should run in a
background job.

The user import file is uploaded to S3 first so the sidekiq workers can
access it.

Once uploaded a job is added that downloads the user import file and
imports it. This job:

- does not retry if any user cannot be imported because of an
  `InvalidSalesforceId`
- retries if there is a problem downloading the file from S3

As there is an need to provide this feature, at this stage log the
steps to Rollbar:

- user imported file uploaded to S3 successfully
- user import background job started
- users imported successfully

Later work will add admin UI around this to make the process more
transparent to users.

Co-authored-by: Laura Porter <laura@dxw.com>